% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_avg_temporal_network_degree}
\alias{calculate_avg_temporal_network_degree}
\title{Calculate average total degree (in + out) in a directed network}
\usage{
calculate_avg_temporal_network_degree(g)
}
\arguments{
\item{g}{A directed igraph object}
}
\value{
Numeric: average number of edges (in + out) per node
}
\description{
Calculate average total degree (in + out) in a directed network
}
