% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_betweenness}
\alias{calculate_betweenness}
\title{Function to calculate time-dependent betweenness centrality}
\usage{
calculate_betweenness(g, shortest_paths_matrix)
}
\description{
Function to calculate time-dependent betweenness centrality
}
