% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Function to calculate Distance (smallest number of nodes, hops, between species)}
\usage{
calculate_distance(g)
}
\description{
Function to calculate Distance (smallest number of nodes, hops, between species)
}
