% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_distance_weights}
\alias{calculate_distance_weights}
\title{Function to calculate Distance (shortest paths with weights representing overlap duration)}
\usage{
calculate_distance_weights(g)
}
\description{
Function to calculate Distance (shortest paths with weights representing overlap duration)
}
