% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_edge_persistence}
\alias{calculate_edge_persistence}
\title{Create a function to calculate edge persistence (pairwise overlap duration)}
\usage{
calculate_edge_persistence(species_data)
}
\description{
Create a function to calculate edge persistence (pairwise overlap duration)
}
