% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_latency}
\alias{calculate_latency}
\title{Function to calculate Latency (time duration, shortest time between species in time-ordered network)}
\usage{
calculate_latency(g)
}
\description{
Function to calculate Latency (time duration, shortest time between species in time-ordered network)
}
