% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_memory_with_latency}
\alias{calculate_memory_with_latency}
\title{The formula uses latency to adjust the memory metric, using correlation between inter-event times}
\usage{
calculate_memory_with_latency(species_data)
}
\description{
The formula uses latency to adjust the memory metric, using correlation between inter-event times
}
