% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_persistence_metrics}
\alias{calculate_persistence_metrics}
\title{Compute average node and edge persistence from the network}
\usage{
calculate_persistence_metrics(g, species_data)
}
\arguments{
\item{g}{An igraph object created from `create_temporal_network`}

\item{species_data}{The original data.frame with species, start_day, and end_day}
}
\value{
A data.frame with two metrics
}
\description{
Compute average node and edge persistence from the network
}
