% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_temporal_network}
\alias{create_temporal_network}
\title{Create a temporal ecological network}
\usage{
create_temporal_network(species_data)
}
\arguments{
\item{species}{A data.frame of nodes}
}
\value{
A networkDynamic object
}
\description{
Create a temporal ecological network
}
