% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inter_event_times}
\alias{inter_event_times}
\title{Calculate inter-event times (duration between end of one species and start of the next)}
\usage{
inter_event_times(species_data)
}
\description{
Calculate inter-event times (duration between end of one species and start of the next)
}
