% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_temporal_network}
\alias{plot_temporal_network}
\title{Plot a snapshot of the temporal network}
\usage{
plot_temporal_network(g)
}
\arguments{
\item{g}{temporal graph}
}
\description{
Plot a snapshot of the temporal network
}
